refactor: extract tier_enforcement module, fix spec-check and arch-check
