+++ src/autopoiesis/tools/exec_tool.py
@@ -15,32 +15,12 @@ from pydantic_ai import RunContext
 from pydantic_ai.messages import ToolReturn
 
 from autopoiesis.infra import exec_registry
-from autopoiesis.infra.command_classifier import Tier, classify
 from autopoiesis.infra.pty_spawn import PtyProcess, read_master, spawn_pty
 from autopoiesis.io_utils import tail_lines
 from autopoiesis.models import AgentDeps
+from autopoiesis.tools.tier_enforcement import enforce_tier
 
 _background_tasks: set[asyncio.Task[None]] = set()
-
-
-def enforce_tier(command: str, approval_unlocked: bool) -> ToolReturn | None:
-    """Check command tier and return a blocked ToolReturn if not permitted."""
-    tier = classify(command)
-    if tier is Tier.BLOCK:
-        return ToolReturn(
-            return_value=f"Blocked: command classified as {tier.value}.",
-            metadata={"blocked": True, "tier": tier.value},
-        )
-    if not approval_unlocked and tier is not Tier.FREE:
-        return ToolReturn(
-            return_value=(
-                f"Approval required: command classified as {tier.value}. "
-                "Unlock approval keys or use Docker backend."
-            ),
-            metadata={"blocked": True, "tier": tier.value},
-        )
-    return None
-
 
 _DANGEROUS_ENV_VARS: frozenset[str] = frozenset(
     {
